requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
django-bootstrap4==3.0.1
django-restframework==0.0.1
djangorestframework==3.12.4
idna==3.2
importlib-metadata==2.1.1
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wincertstore==0.2
zipp==3.6.0
